Extract WhatsApp webhook runtime
